pyproject.toml
[project]
name = "faraalkhata"
version = "0.1.0"
description = "Seasonal Diwali Faraal Order Management System."
readme = "README.md"
authors = [
    { name = "penzulo", email = "bhardesh2005@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.22.0",
    "alembic>=1.17.2",
    "argon2-cffi>=25.1.0",
    "email-validator>=2.3.0",
    "fastapi[standard]>=0.124.4",
    "httpx>=0.28.1",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = [
   "E",
    "F",
    "B",
    "B9",
    "C4",
    "SIM",
    "I",
    "UP",
    "PIE",
    "PGH",
    "PYI",
    "RUF",
]
ignore = [
    "B011",
    "E501",
]

[tool.ruff.format]
docstring-code-format = true

[tool.pyright]
include = [
    ".",
]
exclude = [
    "**/__pycache__",
    "**./*",
    ".venv",
    "build",
    "dist",
    "docs",
    "tests/",
    "**/node_modules",
]
venvPath = "."
venv = ".venv"
pythonVersion = "3.11.13"
pythonPlatform = "Linux"
typeCheckingMode = "standard"
strictListInference = true
strictDictInference = true
strictSetInference = true
reportUnusedImport = false
reportUnusedVariable = false
reportPrivateUsage = false
deprecateTypingAlias = false
reportUnusedClass = false
reportUnusedFunction = false
reportOptionalMemberAccess = true
reportOptionalCall = true
reportOptionalSubscript = true
reportOptionalIterable = true
reportUnknownMemberType = "warning"
reportUnknownVariableType = "warning"
reportUnknownArgumentType = "warning"
reportUnknownParameterType = "warning"
reportImplicitStringConcatenation = "error"
reportInconsistentConstructor = true
